backend/pyproject.toml
[tool.poetry]
name = "guidio-backend"
version = "0.1.0"
description = ""
authors = ["Amar Badnjevic <badnja25@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
SQLAlchemy = "^1.4.46"
starlette = "^0.27.0"
passlib = "^1.7.4"
humps = "^0.2.2"
uvicorn = "^0.22.0"
python-dotenv = "^0.21.1"
alembic = "^1.10.2"
email-validator = "^2.0.0.post2"
python-jose = "^3.3.0"
psycopg2 = "^2.9.6"
python-multipart = "^0.0.5"
bcrypt = "^4.0.1"
fastapi-mail = "^1.4.1"

[tool.poetry.scripts]
guidio = "src.main:main"
migrate = "migrate:run_alembic_upgrade"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
